+++ common/src/main/java/io/ejekta/bountiful/Bountiful.kt
@@ -5,7 +5,6 @@ import io.ejekta.bountiful.bounty.types.IBountyType
 import io.ejekta.kambrik.Kambrik
 import io.ejekta.kambrik.serial.serializers.IdentitySer
 import kotlinx.serialization.UseSerializers
-import net.fabricmc.api.ModInitializer
 import net.minecraft.registry.Registry
 import net.minecraft.registry.RegistryKey
 import net.minecraft.util.Identifier