+++ common/src/main/java/io/ejekta/bountiful/bounty/DecreeData.kt
@@ -1,7 +1,6 @@
 package io.ejekta.bountiful.bounty
 
 import io.ejekta.bountiful.Bountiful
-import io.ejekta.bountiful.config.BountifulIO
 import io.ejekta.bountiful.content.BountifulContent
 import io.ejekta.kambrik.serial.ItemDataJson
 import kotlinx.serialization.Serializable