+++ common/src/main/java/io/ejekta/bountiful/bounty/types/builtin/BountyTypeCommand.kt
@@ -1,10 +1,8 @@
 package io.ejekta.bountiful.bounty.types.builtin
 
-import io.ejekta.bountiful.Bountiful
 import io.ejekta.bountiful.bounty.BountyDataEntry
 import io.ejekta.bountiful.bounty.types.IBountyReward
 import io.ejekta.bountiful.data.PoolEntry
-import io.ejekta.kambrik.text.textLiteral
 import net.minecraft.entity.player.PlayerEntity
 import net.minecraft.server.MinecraftServer
 import net.minecraft.text.MutableText