+++ common/src/main/java/io/ejekta/bountiful/bridge/BountifulSharedApi.kt
@@ -17,7 +17,6 @@ import net.minecraft.client.item.ModelPredicateProviderRegistry
 import net.minecraft.entity.Entity
 import net.minecraft.entity.LivingEntity
 import net.minecraft.entity.passive.TameableEntity
-import net.minecraft.entity.player.PlayerEntity
 import net.minecraft.server.network.ServerPlayerEntity
 import net.minecraft.server.world.ServerWorld