+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) as closely as it can.
 
-## [7.0.0] for 1.20.(4?) - WIP
+## [7.0.0] for 1.20.(4?) - Not Yet Released
 
 This is largely a Quality Of Life update, making it easier for modpack makers to configure Bountiful. As such,
 old config files and datapacks that overwrote existing Bountiful data may not function correctly in this version.